+++ addon/components/flexberry-groupedit.js
@@ -14,6 +14,7 @@ import FlexberryBaseComponent from './flexberry-base-component';
 import Information from 'ember-flexberry-data/utils/information';
 import { translationMacro as t } from 'ember-i18n';
 import getProjectionByName from '../utils/get-projection-by-name';
+import sortRecords from '../utils/sorting-function';
 
 /**
   Component for create, edit and delete detail objects.
@@ -667,20 +668,20 @@ export default FlexberryBaseComponent.extend({
       for (let i = 0; i < sorting.length; i++) {
         let sort = sorting[i];
         if (i === 0) {
-          records = this.sortRecords(records, sort, 0, records.length - 1);
+          records = sortRecords(records, sort, 0, records.length - 1);
         } else {
           let index = 0;
           for (let j = 1; j < records.length; j++) {
             for (let sortIndex = 0; sortIndex <  i; sortIndex++) {
               if (records.objectAt(j).get(sorting[sortIndex].propName) !== records.objectAt(j - 1).get(sorting[sortIndex].propName)) {
-                records = this.sortRecords(records, sort, index, j - 1);
+                records = sortRecords(records, sort, index, j - 1);
                 index = j;
                 break;
               }
             }
           }
 
-          records = this.sortRecords(records, sort, index, records.length - 1);
+          records = sortRecords(records, sort, index, records.length - 1);
         }
       }
 
@@ -689,81 +690,6 @@ export default FlexberryBaseComponent.extend({
 
     let componentName = this.get('componentName');
     this.get('_groupEditEventsService').geSortApplyTrigger(componentName, this.get('sorting'));
-  },
-
-  /**
-    Client-side sorting for groupEdit content.
-
-    @method sortRecords
-    @param {Array} records Records for sorting.
-    @param {Object} sortDef Sorting definition.
-    @param {Int} start First index in records.
-    @param {Int} end Last index in records.
-    @return {Array} Sorted records.
-  */
-  sortRecords(records, sortDef, start, end) {
-    let recordsSort = records;
-    if (start >= end) {
-      return recordsSort;
-    }
-
-    // Form hash array (there can be different observers on recordsSort changing, so it is better to minimize such changes).
-    let hashArray = [];
-    for (let i = start; i <= end; i++) {
-      let currentRecord = recordsSort.objectAt(i);
-      let currentHash = currentRecord.get(sortDef.attributePath || sortDef.propName);
-      let hashStructure = {
-        record: currentRecord,
-        hash: currentHash
-      };
-
-      hashArray.push(hashStructure);
-    }
-
-    let hashArrayLength = hashArray.length;
-
-    // Compare record with number koef1 and koef2.
-    // It returns true if records should be exchanged.
-    let condition = function(koef1, koef2) {
-      let firstProp = hashArray[koef1].hash;
-      let secondProp = hashArray[koef2].hash;
-      if (sortDef.direction === 'asc') {
-        return isNone(secondProp) && !isNone(firstProp) ? true : firstProp > secondProp;
-      }
-
-      if (sortDef.direction === 'desc') {
-        return !isNone(secondProp) && isNone(firstProp) ? true : firstProp < secondProp;
-      }
-
-      return false;
-    };
-
-    // Sort with minimum exchanges.
-    for(let i = 0; i < hashArrayLength; i++) {
-      // Find minimum in right not sorted part.
-      let min = i;
-      for(let j = i + 1; j < hashArrayLength; j++) {
-        if(condition(min, j)) {
-          min = j;
-        }
-      }
-      if (min != i) {
-        // Exchange current with minimum.
-        let tmp = hashArray[i];
-        hashArray[i] = hashArray[min];
-        hashArray[min] = tmp;
-      }
-    }
-
-    // Remove unsorted part.
-    recordsSort.removeAt(start, end - start + 1);
-
-    // Insert sorted elements.
-    for (let i = start; i <= end; i++) {
-      recordsSort.insertAt(i, hashArray[i-start].record);
-    }
-
-    return recordsSort;
   },
 
   init() {