+++ addon/mixins/flexberry-objectlistview-hierarchical-route.js
@@ -7,8 +7,10 @@ import RSVP from 'rsvp';
 import $ from 'jquery';
 import { isNone } from '@ember/utils';
 import { set } from '@ember/object';
-import { A } from '@ember/array';
+import { A, isArray } from '@ember/array';
 import Builder from 'ember-flexberry-data/query/builder';
+import sortRecords from '../utils/sorting-function';
+import deserializeSortingParam from '../utils/deserialize-sorting-param';
 
 /**
   Mixin for [Route](https://www.emberjs.com/api/ember/release/classes/Route) to support hierarchical mode into {{#crossLink "FlexberryObjectlistviewComponent"}}{{/crossLink}}.
@@ -32,6 +34,7 @@ export default Mixin.create({
     loadRecordsById(id, target, property, firstRunMode, recordParams) {
       let params = recordParams || {};
       let hierarchicalAttribute = params.hierarchicalAttribute || this.controllerFor(this.routeName).get('hierarchicalAttribute');
+      let sorting = deserializeSortingParam(params.sort || this.controllerFor(this.routeName).get('sort'));
       let modelName = params.modelName || this.get('modelName');
 
       if (firstRunMode) {
@@ -40,6 +43,13 @@ export default Mixin.create({
           .from(modelName)
           .selectByProjection(projectionName)
           .where(hierarchicalAttribute, 'eq', id);
+
+        if (isArray(sorting)) {
+          sorting = sorting.filter(i => i.direction !== 'none').map(i => `${i.propName} ${i.direction}`).join(',');
+          if (sorting) {
+            builder.orderBy(sorting);
+          }
+        }
 
         set(target, property, this.store.query(modelName, builder.build()));
       } else {
@@ -59,6 +69,26 @@ export default Mixin.create({
           this.send('loadRecordsById', id, target, property, true, recordParams);
         } else {
           let recordsArrayinPromise = new RSVP.Promise((resolve) => {
+            for (let i = 0; i < sorting.length; i++) {
+              let sort = sorting[i];
+              if (i === 0) {
+                sortRecordsArray = sortRecords(sortRecordsArray, sort, 0, sortRecordsArray.length - 1);
+              } else {
+                let index = 0;
+                for (let j = 1; j < sortRecordsArray.length; j++) {
+                  for (let sortIndex = 0; sortIndex <  i; sortIndex++) {
+                    if (sortRecordsArray.objectAt(j).get(sorting[sortIndex].propName) !== sortRecordsArray.objectAt(j - 1).get(sorting[sortIndex].propName)) {
+                      sortRecordsArray = sortRecords(sortRecordsArray, sort, index, j - 1);
+                      index = j;
+                      break;
+                    }
+                  }
+                }
+      
+                sortRecordsArray = sortRecords(sortRecordsArray, sort, index, sortRecordsArray.length - 1);
+              }
+            }
+
             resolve(sortRecordsArray);
           });